% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNat_IDs.R
\name{iNat_IDs}
\alias{iNat_IDs}
\title{iNaturalist Identifications Import}
\usage{
iNat_IDs(
  per_page = 200,
  order = "desc",
  order_by = NULL,
  current_taxon = NULL,
  own_observation = NULL,
  is_change = NULL,
  taxon_active = NULL,
  observation_taxon_active = NULL,
  id = NULL,
  rank = NULL,
  observation_rank = NULL,
  user_id = NULL,
  user_login = NULL,
  current = NULL,
  category = NULL,
  place_id = NULL,
  quality_grade = NULL,
  taxon_id = NULL,
  observation_taxon_id = NULL,
  iconic_taxon_id = NULL,
  observation_iconic_taxon_id = NULL,
  lrank = NULL,
  hrank = NULL,
  observation_lrank = NULL,
  observation_hrank = NULL,
  without_taxon_id = NULL,
  d1 = NULL,
  d2 = NULL,
  observation_created_d1 = NULL,
  observation_created_d2 = NULL,
  observed_d1 = NULL,
  observed_d2 = NULL,
  id_above = NULL,
  id_below = NULL,
  page = NULL
)
}
\arguments{
\item{per_page}{Results requested per API request page. Cannot exceed 200}

\item{order}{Query sort order. "desc" (default) for descending and "asc" for ascending}

\item{order_by}{Sorting field for order paramter. Options are "created_at" (default) and "id"}

\item{current_taxon}{Whether or not the current taxon is identical to the observation's taxon: "true" or "false"}

\item{own_observation}{Whether or not the identification is for an observation of the identifier: "true" or "false"}

\item{is_change}{Whetheror not the identification is the automatic result of a taxonomic change in the iNaturalist database: "true" or "false"}

\item{taxon_active}{Whether or not the taxon ID is active in the iNaturalist database: "true" or "false"}

\item{observation_taxon_active}{Whether or not the taxon of the observation is active in the iNaturalist database: "true" or "false"}

\item{id}{Fetch a specific identification ID number}

\item{rank}{Specify the ID taxonomic rank: "kingdom", "phylum", "subphylum", "superclass", "class", "subclass", "superorder", "order", "suborder", "infraorder", "superfamily", "epifamily", "family", "subfamily", "supertribe", "tribe", "subtribe", "genus", "genushybrid", "species", "hybrid", "subspecies", "variety", "form",}

\item{observation_rank}{Specify the observation taxonomic rank: "kingdom", "phylum", "subphylum", "superclass", "class", "subclass", "superorder", "order", "suborder", "infraorder", "superfamily", "epifamily", "family", "subfamily", "supertribe", "tribe", "subtribe", "genus", "genushybrid", "species", "hybrid", "subspecies", "variety", "form",}

\item{user_id}{Fetch identifications from user ID number}

\item{user_login}{Fetch identifications by "[user name]"}

\item{current}{Fetch identifications that are the current, active ones made by a user (i.e., have not been retracted): "true" (default) or "false"}

\item{category}{Identification type: "improving" (brings the observation ID to a more specific level and has community agreement), "supporting" (confirms an existing level of observation ID), "leading" (brings the observation IF to a more specific level but has not yet been confirmed by the community), and "maverick" (an ID that conflicts with the community consensus ID for an observation; not necessarily a wrong ID)}

\item{place_id}{Fetch identifications based on place ID}

\item{quality_grade}{Current quality grade of the observation being ID'ed: "casual", "needs_id", "research"}

\item{taxon_id}{Fetch identifications that are of taxon ID and descendants}

\item{observation_taxon_id}{Fetch identifications that are for an observation of a taxon and descendants}

\item{iconic_taxon_id}{Fetch identifications that are of an iconic taxon ID}

\item{observation_iconic_taxon_id}{Fetch identifications that are for an observation of an iconic taxon}

\item{lrank}{Fetch identifications with this taxon rank or higher. Example options are: "kingdom", "phylum", "order", "family", "genus", "species"}

\item{hrank}{Fetch identifications with this taxon rank or lower. Example options are: "kingdom", "phylum", "order", "family", "genus", "species"}

\item{observation_lrank}{Fetch identifications for observations with this taxon rank or higher. Example options are: "kingdom", "phylum", "order", "family", "genus", "species"}

\item{observation_hrank}{Fetch identifications for observations with this taxon rank or lower. Example options are: "kingdom", "phylum", "order", "family", "genus", "species"}

\item{without_taxon_id}{Fetch identifications that are not of a taxon or its descendants}

\item{d1}{Fetch identifications on or after a specific date: "YYYY-MM-DD"}

\item{d2}{Fetch identifications on or before a specific date: "YYYY-MM-DD"}

\item{observation_created_d1}{Fetch identifications for observations created on or after a specific date: "YYYY-MM-DD"}

\item{observation_created_d2}{Fetch identifications for observations created on or before a specific date: "YYYY-MM-DD"}

\item{observed_d1}{Fetch identifications for observations that were observed on or after a specific date: "YYYY-MM-DD"}

\item{observed_d2}{Fetch identifications for observations that were observed on or before a specific date: "YYYY-MM-DD"}

\item{id_above}{Fetch identifications above this ID number}

\item{id_below}{Fetch identifications below this ID number}
}
\description{
Function for pulling identification data from the iNaturalist API
}
\examples{

Fetch species counts for Baker Woodlot at Michigan State University:
df <- iNat_IDs(place_id=136412)

Fetch species counts of insects for a specific user at a specific location (Baker Woodlot):
df <- iNat_IDs(user_id="hanly", taxon_id=47158, place_id=136412)
}
